Reorder types files
