+++ app/support/types/app-ctx.ts
@@ -1,0 +1,8 @@
+import { type DefaultState, type ParameterizedContext } from 'koa';
+import { type Config } from 'config';
+
+import PubsubListener from '../../pubsub-listener';
+
+// Controller context
+export type AppContext = { config: Config; port: number; pubsub: PubsubListener };
+export type Ctx<State = DefaultState> = ParameterizedContext<State, AppContext>;
