+++ app/support/types/branded.ts
@@ -1,0 +1,5 @@
+import { Branded } from './helpers';
+
+export type UUID = Branded<string, 'uuid'>;
+export type IPAddr = Branded<string, 'IP address'>;
+export type ISO8601DurationString = Branded<string, 'ISO 8601 Duration'>;
